pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "db-agent-app"
version = "0.1.0"
description = "Database Analysis Agent - AI-powered natural language database querying"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "cryptography==41.0.7",
    "dash==2.17.0",
    "dash-bootstrap-components==1.5.0",
    "dash-extensions==1.0.4",
    "flask==2.3.3",
    "gunicorn==21.2.0",
    "lancedb>=0.2.0",
    "langchain>=0.3.0",
    "langchain-community==0.3.16",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "litellm>=1.0.0",
    "matplotlib>=3.8.0",
    "numpy==1.26.2",
    "openai==1.102.0",
    "pandas==2.1.4",
    "plotly==5.17.0",
    "psycopg2-binary==2.9.9",
    "pymysql==1.1.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "seaborn>=0.13.0",
    "sqlalchemy==2.0.23",
    "sqlfluff>=2.3.0",
    "sqlglot>=20.0.0",
]

[tool.setuptools.packages.find]
include = ["database*", "services*"]
